Test the ECS infra
